+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.6
+
+- Info about attachments in messages view
+- Preview of images in messages view (only with terminals that support it, eg Kitty, iTerm2)
+
 ## v0.5.2
 
 - Compose multi line messages